+++ src/challenges/js/30DaysJs-25/src/Card.class.js
@@ -1,0 +1,21 @@
+import { Pay } from "./Pay.class.js";
+
+export class Card extends Pay {
+  constructor(cardNumber) {
+    super();
+    this.cardNumber = cardNumber;
+  }
+
+  makePay(quantity) {
+    if (this.cardNumber.length !== 16) {
+      throw new Error("Tarjeta invalida");
+    }
+
+    const lastCardNumbers = this.cardNumber.toString().slice(-4);
+
+    return {
+      ...super.makePay(quantity),
+      lastCardNumbers,
+    };
+  }
+}
