+++ src/challenges/js/30DaysJs-25/src/Cash.class.js
@@ -1,0 +1,7 @@
+import { Pay } from "./Pay.class.js";
+
+export class Cash extends Pay {
+  constructor() {
+    super();
+  }
+}
