+++ src/challenges/js/30DaysJs-25/src/Pay.class.js
@@ -1,0 +1,10 @@
+export class Pay {
+  constructor() {}
+
+  makePay(quantity) {
+    return {
+      realized: true,
+      quantity,
+    };
+  }
+}
