+++ src/challenges/js/30DaysJs-25/src/PayPal.class.js
@@ -1,0 +1,16 @@
+import { Pay } from "./Pay.class.js";
+
+export class PayPal extends Pay {
+  constructor(email) {
+    super();
+    this.email = email;
+  }
+
+  makePay(quantity) {
+    return {
+      ...super.makePay(quantity),
+      platform: "PayPal",
+      email: this.email,
+    };
+  }
+}
